+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Check docs generation with LDoc.
 - Add `--repeat-group` (`-R`) option to run tests in a circle within the group.
+- Add `flaky` test mark.
 
 ## 0.5.7