+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Make the internal server_instance.lua script compatible with Tarantool 
+  versions < 2.8.0-139-g357f15517.
 - Extend `server.lua` API:
   * Update parameters of the `Server:new()` function:
     - The `alias` parameter defaults to 'server'.